backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
pydantic-settings==2.12.0
langchain==1.2.8
langgraph==1.0.7
langchain-groq==1.1.1
python-dotenv==1.2.1
python-multipart==0.0.22
aiohttp==3.13.3
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
motor==3.7.1
pymongo==4.12.0
beanie==1.29.0
python-jose[cryptography]==3.5.0
aiofiles==24.1.0
email-validator==2.3.0
cryptography>=42.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4